requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
greenlet==3.3.2
idna==3.11
psycopg2-binary==2.9.9
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.47
typing_extensions==4.15.0
urllib3==2.6.3